requirements.txt
openai
tiktoken
scipy
gspread
google-auth
